Move gasp table building to InstructionCompiler
